Use internal data
Aggregating data from qmongrdata takes forever,
and slows down the starting of the app (see issue #112).
In this commit, the data `aggr_data` is created by running
the function app_data() beforehand, and is used directly by the app.

This is just a quick fix and not a final solution, because the app now
does not depend on external data.